Adjust controller file structure to match logical
